+++ delphi-frontend/src/test/resources/au/com/integradev/delphi/symbol/properties/DefaultProperties.pas
@@ -1,0 +1,15 @@
+unit DefaultProperties;
+
+interface
+
+const
+  C_Bar = 123;
+
+type
+  TFoo = class
+    property Bar: Integer default C_Bar;
+  end;
+
+implementation
+
+end.