+++ delphi-checks/src/main/java/au/com/integradev/delphi/checks/UnusedPropertyCheck.java
@@ -56,6 +56,10 @@ public class UnusedPropertyCheck extends DelphiCheck {
       return false;
     }
 
+    if (!declaration.getImplementedTypes().isEmpty()) {
+      return false;
+    }
+
     if (excludeApi && declaration.isPublic() && !declaration.isImplementationDeclaration()) {
       return false;
     }