+++ CHANGELOG.md
@@ -38,7 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Include "found" and "expected" values for issues messages in `LowercaseKeyword`.
+- Include "found" and "expected" values for issue messages in `LowercaseKeyword`.
 - Exclude `J` and `K` by default in `ShortIdentifier`.
 - Improve compiler directive parsing in cases where the directive name is not followed by whitespace.